Make the antlr demo higher quality
