--- antlr_demo/magic.py
@@ -1,4 +1,0 @@
-import sys 
-
-print("jeffrey")
-print("fergus", file=sys.stderr)
